Add: Persistencia perfil y usuario + Nombres consistentes en controladores y servicios

+++ backend/src/main/java/app/repositories/RepositorioPerfiles.java
@@ -12,7 +12,7 @@ public interface RepositorioPerfiles {
 
     List<Perfil> buscarTodos();
 
-    int contar();
+    long contar();
 
     List<Perfil> buscarPorFiguritaFaltante(Figurita figurita);
